Updates and Bug Fixes
updated the functionality to make the manual threshold an option for users.

updated the functionality of the file creator to make it add a number to a file if it already exists

fixed a bug where during the merge process I missed a call of the anomaly detection class in the anomaly report function causing it to reset the initial arguments it got fed.